Cleanup: move getRebalance() into rebalance.go
